Change 'dividen' column type from integer to float in balances migration

+++ database/migrations/_create_balances.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('nama');
             $table->integer('saldo');
             $table->string('tipe');
-            $table->integer('dividen')->nullable();
+            $table->float('dividen', 8, 2)->nullable();
             $table->timestamps();
         });
     }
